implementing dto class and including bean validations, configuring datasource on properties

+++ src/main/kotlin/br/com/service/AccountBankService.kt
@@ -1,6 +1,7 @@
 package br.com.service
 
 import br.com.domain.Account
+import java.math.BigDecimal
 
 /**
  * Interface defines account bank contract operations.
@@ -9,9 +10,9 @@ interface AccountBankService {
 
     fun persist(account: Account): Account
 
-    fun getBalance(account: Account): Double
+    fun getBalance(account: Account): BigDecimal
 
     fun getCashAccount(account: Account): Account
 
-    fun getOverdrawn(account: Account): Double
+    fun getOverdrawn(account: Account): BigDecimal
 }
